requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.4.3=pypi_0
aiohttp=3.10.10=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.7.0=pypi_0
anyascii=0.3.2=pypi_0
anyio=4.6.2.post1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=24.2.0=pypi_0
audioread=3.0.1=pypi_0
babel=2.16.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blis=0.7.11=pypi_0
ca-certificates=2024.9.24=hca03da5_0
catalogue=2.0.10=pypi_0
certifi=2024.8.30=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.0=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.20.0=pypi_0
confection=0.1.5=pypi_0
contourpy=1.3.0=pypi_0
coqpit=0.0.17=pypi_0
coqui-tts=0.24.2=pypi_0
coqui-tts-trainer=0.1.4=pypi_0
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
cython=3.0.11=pypi_0
dateparser=1.1.8=pypi_0
decorator=5.1.1=pypi_0
distro=1.9.0=pypi_0
docopt=0.6.2=pypi_0
ebooklib=0.18=pypi_0
edge-tts=6.1.12=pypi_0
einops=0.8.0=pypi_0
encodec=0.1.1=pypi_0
exceptiongroup=1.2.2=pypi_0
filelock=3.16.1=pypi_0
fonttools=4.54.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.9.0=pypi_0
grpcio=1.67.0=pypi_0
gruut=2.4.0=pypi_0
gruut-ipa=0.13.0=pypi_0
gruut-lang-de=2.0.1=pypi_0
gruut-lang-en=2.0.1=pypi_0
gruut-lang-es=2.0.1=pypi_0
gruut-lang-fr=2.0.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.6=pypi_0
httpx=0.27.2=pypi_0
huggingface-hub=0.26.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
inflect=7.4.0=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
jsonlines=1.2.0=pypi_0
kiwisolver=1.4.7=pypi_0
langcodes=3.4.1=pypi_0
language-data=1.2.0=pypi_0
lazy-loader=0.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
librosa=0.10.2.post1=pypi_0
llvmlite=0.43.0=pypi_0
lxml=5.3.0=pypi_0
marisa-trie=1.2.1=pypi_0
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.9.2=pypi_0
mdurl=0.1.2=pypi_0
more-itertools=10.5.0=pypi_0
mpmath=1.3.0=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
murmurhash=1.0.10=pypi_0
mutagen=1.47.0=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.2.1=pypi_0
num2words=0.5.13=pypi_0
numba=0.60.0=pypi_0
numpy=1.26.4=pypi_0
openai=1.35.7=pypi_0
openssl=3.0.15=h80987f9_0
packaging=24.1=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py39hca03da5_0
platformdirs=4.3.6=pypi_0
pooch=1.8.2=pypi_0
preshed=3.0.9=pypi_0
propcache=0.2.0=pypi_0
protobuf=5.28.2=pypi_0
psutil=6.1.0=pypi_0
pycparser=2.22=pypi_0
pydantic=2.9.2=pypi_0
pydantic-core=2.23.4=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pysbd=0.3.4=pypi_0
python=3.9.20=hb885b13_1
python-crfsuite=0.9.11=pypi_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2024.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h1a28f6b_0
regex=2024.9.11=pypi_0
requests=2.32.3=pypi_0
rich=13.9.2=pypi_0
safetensors=0.4.5=pypi_0
scikit-learn=1.5.2=pypi_0
scipy=1.13.1=pypi_0
setuptools=75.1.0=py39hca03da5_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smart-open=7.0.5=pypi_0
sniffio=1.3.1=pypi_0
socksio=1.0.0=pypi_0
soundfile=0.12.1=pypi_0
soupsieve=2.6=pypi_0
soxr=0.5.0.post1=pypi_0
spacy=3.7.5=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.45.3=h80987f9_0
srsly=2.4.8=pypi_0
sudachidict-core=20240716=pypi_0
sudachipy=0.6.8=pypi_0
sympy=1.13.1=pypi_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thinc=8.2.5=pypi_0
threadpoolctl=3.5.0=pypi_0
tk=8.6.14=h6ba3021_0
tokenizers=0.19.1=pypi_0
torch=2.5.0=pypi_0
torchaudio=2.5.0=pypi_0
tqdm=4.66.5=pypi_0
transformers=4.42.4=pypi_0
typeguard=4.3.0=pypi_0
typer=0.12.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
tzlocal=5.2=pypi_0
urllib3=2.2.3=pypi_0
wasabi=1.1.3=pypi_0
weasel=0.4.1=pypi_0
werkzeug=3.0.4=pypi_0
wheel=0.44.0=py39hca03da5_0
wrapt=1.16.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.15.5=pypi_0
zipp=3.20.2=pypi_0
zlib=1.2.13=h18a0788_1
